Rename to io_stream.h for consistency
