+++ tensorflow_io/avro/kernels/avro_kernels.cc
@@ -14,8 +14,8 @@ limitations under the License.
 ==============================================================================*/
 
 #include "tensorflow/core/framework/op_kernel.h"
-#include "tensorflow_io/core/kernels/stream.h"
 #include "tensorflow_io/core/kernels/io_interface.h"
+#include "tensorflow_io/core/kernels/io_stream.h"
 #include "api/DataFile.hh"
 #include "api/Compiler.hh"
 #include "api/Generic.hh"
