+++ tensorflow_io/core/kernels/io_stream.h
