+++ tensorflow_io/hdf5/kernels/hdf5_kernels.cc
@@ -15,7 +15,7 @@ limitations under the License.
 
 #include "tensorflow/core/framework/op_kernel.h"
 #include "tensorflow_io/core/kernels/io_interface.h"
-#include "tensorflow_io/core/kernels/stream.h"
+#include "tensorflow_io/core/kernels/io_stream.h"
 
 #include <hdf5.h>
 #include <hdf5_hl.h>
