+++ tensorflow_io/text/kernels/csv_kernels.cc
@@ -14,10 +14,9 @@ limitations under the License.
 ==============================================================================*/
 
 #include "tensorflow/core/framework/op_kernel.h"
-#include "tensorflow_io/core/kernels/stream.h"
 #include "tensorflow/core/lib/io/buffered_inputstream.h"
 #include "tensorflow_io/core/kernels/io_interface.h"
-#include "tensorflow_io/core/kernels/stream.h"
+#include "tensorflow_io/core/kernels/io_stream.h"
 #include "arrow/memory_pool.h"
 #include "arrow/csv/reader.h"
 #include "arrow/table.h"
