+++ tensorflow_io/text/kernels/text_kernels.cc
@@ -14,8 +14,8 @@ limitations under the License.
 ==============================================================================*/
 
 #include "tensorflow/core/framework/op_kernel.h"
+#include "tensorflow_io/core/kernels/io_stream.h"
 #include "tensorflow/core/lib/io/buffered_inputstream.h"
-#include "tensorflow_io/core/kernels/stream.h"
 
 namespace tensorflow {
 namespace data {
